Separated old zign functionality to a different file
