+++ setup.py
@@ -51,8 +51,8 @@ CLASSIFIERS = [
 
 CONSOLE_SCRIPTS = [
     # TODO: deprecate "zign"
-    'zign = zign.cli:main',
-    'ztoken = zign.cli_ztoken:main'
+    'ztoken = zign.cli:main',
+    'zign = zign.cli_zign:main'
 ]
